feat: adding Ubuntu support
This adds support for Ubuntu debs. It also adds some corresponding smoke
tests, but I'm not 100% sure I did that correctly.

This "Works On My Machine(tm)", but I've not extensively tested it on
anything other than Ubuntu 22.04 LTS. There may be a more general
approach to handle the differences between Debian and Ubuntu apt
packaging topologies, but I saw that you are exploring a native Skylark
port (https://github.com/bazel-contrib/rules_debian_packages/issues/30)
so I didn't think the refactor would be worth the time investment.

This also fixes a bug which I hit where a debian package is not
`xz`-compressed, by adding support for `.tar.zst` in the genrule.

This is done by modifying the buildfile contents in
`repository.packages.tmpl` to pass the extracted data archive as a
single-entry filegroup, then modifying the genrule to iterate over the
inputs and decompress them. This will have the side effect of
potentially producing anther label, but offhand I can't think of a
situation where that will cause issues.

The one other failure mode is if a `.deb` is somehow malformed in such a
way that it contains *two* `data.tar.*` files, in which case it just
extracts the first it sees and continues.